Update for Winter 2020 Jams
Changed date. Pointed tickets to Esports eventbrite (in two places) - if we build our own we need to change this.
Changed location and updated link to google for directions.

Still need to change colors, but maybe for the last time to a black and white color scheme.

There are probably other changes that need to be made.